requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.2
django-debug-toolbar==3.7.0
django-stubs==1.12.0
django-stubs-ext==0.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djangorestframework-stubs==1.7.0
drf-spectacular==0.24.2
drf-spectacular-sidecar==2022.11.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.0
MarkupSafe==2.1.1
mypy==0.982
mypy-extensions==0.4.3
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.2
pytz==2022.5
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.3
tomli==2.0.1
types-Markdown==3.4.2.1
types-pytz==2022.5.0.0
types-PyYAML==6.0.12.1
types-requests==2.28.11.2
types-urllib3==1.26.25.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
